pyproject.toml
[tool.poetry]
name = "ai-agent-system"
version = "0.1.0"
description = "Sistema modular de agentes IA basado en MCP"
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.103.1"
uvicorn = "^0.23.2"
pydantic = "^2.3.0"
chromadb = "^0.4.13"
pymongo = "^4.5.0"
python-dotenv = "^1.0.0"
httpx = "^0.25.0"
pyyaml = "^6.0.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.2"
black = "^23.9.1"
isort = "^5.12.0"
flake8 = "^6.1.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api" 